create crud user

+++ app/controllers/Backend/MovieController.php
@@ -3,9 +3,17 @@ require_once 'app/core/Controller.php';
 require_once 'app/models/Movie.php';
 require_once 'app/models/User.php';
 require_once 'app/models/Category.php';
+require_once 'app/models/Video.php';
 
 class MovieController extends Controller
 {
+    public function __construct()
+    {
+        if (!$this->isAuth()) {
+            return $this->redirect(URL . '/admin/login');
+        }
+    }
+
     public function index()
     {
         $model = new Movie();
@@ -15,7 +23,7 @@ class MovieController extends Controller
         foreach ($movies as $movie) {
             $userModel = new User();
             $user = $userModel->find($movie->user_id);
-            $movie->user_id = $user->name;
+            $movie->user = $user->name;
 
             $categoryModel = new Category();
 
@@ -48,6 +56,11 @@ class MovieController extends Controller
         $data = $_POST;
 
         try {
+            if ($data['name'] == ''){
+                $_SESSION['error']['name'] = "Vui lòng nhập tên phim";
+                return $this->redirect($_SERVER['HTTP_REFERER']);
+            }
+
             $data['user_id'] = $_SESSION['auth']['id'];
 
             $data['slug'] = to_slug($data['name']) . strtotime(date('Y-m-d H:i:s'));
@@ -60,6 +73,7 @@ class MovieController extends Controller
 
             $success = $movie->create($data);
             if ($success) {
+                $_SESSION['success'] = "Tạo mới thành công";
                 return $this->redirect(URL . '/admin/movies');
             }
 
@@ -70,7 +84,14 @@ class MovieController extends Controller
 
     public function show($id)
     {
+        $model = new Movie();
 
+        $movie = $model->find($id);
+
+        if ($movie->type == 1)
+            $this->redirect(URL.'/tv-series/'.$movie->slug);
+        else
+            $this->redirect(URL.'/movies/'.$movie->slug);
     }
 
     public function edit($id)
@@ -80,6 +101,11 @@ class MovieController extends Controller
         $movie = $model->find($id);
         $model = new Category();
         $categories = $model->all();
+
+        if ($movie->user_id != $_SESSION['auth']['id'] && $_SESSION['auth']['role'] != 1) {
+            echo 'ERROR - 403';
+            exit();
+        }
 
         return $this->views('movies/update', [
             'movie' => $movie,
@@ -92,16 +118,28 @@ class MovieController extends Controller
         $data = $_POST;
 
         try {
+            if ($data['name'] == ''){
+                $_SESSION['error']['name'] = "Vui lòng nhập tên phim";
+                return $this->redirect($_SERVER['HTTP_REFERER']);
+            }
 
             if ($_FILES['image']['name'] != "")
                 $data['image'] = file_upload('/publics/storage/image', 'image', 5000000, array('JPG', 'png'));
 
             $data['slug'] = to_slug($data['name']) . strtotime(date('Y-m-d H:i:s'));
 
-            $movie = new Movie();
+            $model = new Movie();
 
-            $success = $movie->update($id, $data);
+            $movie = $model->find($id);
+
+            if ($movie->user_id != $_SESSION['auth']['id'] && $_SESSION['auth']['role'] != 1) {
+                echo 'ERROR - 403';
+                exit();
+            }
+
+            $success = $model->update($id, $data);
             if ($success) {
+                $_SESSION['success'] = "Cập nhật thành công";
                 return $this->redirect(URL . '/admin/movies');
             }
 
@@ -114,8 +152,25 @@ class MovieController extends Controller
     {
         $model = new Movie();
 
-        $model->delete($id);
+        $movie = $model->find($id);
 
+        if ($movie->user_id != $_SESSION['auth']['id'] && $_SESSION['auth']['role'] != 1) {
+            echo 'ERROR - 403';
+            exit();
+        }
+
+        $modelVideo = new Video();
+        $videos = $modelVideo->where('movie_id','=',$movie->id)->get();
+
+        foreach ($videos as $video){
+            $data = [
+              'movie_id'=>null
+            ];
+            $modelVideo->update($video->id,$data);
+        }
+
+        $model->delete($id);
+        $_SESSION['success'] = "Xóa thành công";
         return $this->redirect(URL . '/admin/movies');
     }
